loops_api/pyproject.toml
[project]
name = "loops_api"
version = "0.1.0"
description = "API for analyzing and predicting missing loops in protein structures"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.81",
    "requests>=2.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.hatch.build.targets.wheel]
packages = ["loops_api"] 